move provisioning-info cm creation to skr config state
