+++ internal/controller/runtime/fsm/runtime_fsm_create_shoot.go
@@ -65,12 +65,6 @@ func sFnCreateShoot(ctx context.Context, m *fsm, s *systemState) (stateFn, *ctrl
 		}
 	}
 
-	skrDetailsErr:= createKymaProvisioningInfoCM(ctx, m, s)
-	if skrDetailsErr != nil {
-		return handleConfigMapCreationError(m, skrDetailsErr, s)
-	}
-	m.log.V(log_level.DEBUG).Info("kyma-provisioning-info config map is created")
-
 	data, err := m.AuditLogging.GetAuditLogData(
 		s.instance.Spec.Shoot.Provider.Type,
 		s.instance.Spec.Shoot.Region)
